# src/swagger/swagger.yaml
basePath: /
definitions:
  api.ChatArchiveRequest:
    properties:
      archive:
        description: 'Required: true to archive, false to unarchive'
        type: boolean
      chatid:
        description: 'Required: Chat to archive/unarchive'
        type: string
    type: object
  api.ChatPresenceRequest:
    properties:
      chatid:
        description: 'Required: Chat to show typing in'
        type: string
      duration:
        description: 'Optional: Auto-stop after duration (ms)'
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappChatPresenceType'
        description: Text or audio
    type: object
  api.ChatReadRequest:
    properties:
      chatid:
        description: 'Required: Chat to mark as read/unread'
        type: string
    type: object
...
  models.QpResponse:
    properties:
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      status:
        type: string
      success:
        type: boolean
    type: object
...
  whatsapp.WhatsappChatPresenceType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - WhatsappChatPresenceTypePaused
    - WhatsappChatPresenceTypeText
    - WhatsappChatPresenceTypeAudio
...
info:
  contact:
    name: QuePasa Support
    url: https://github.com/nocodeleaks/quepasa
  description: QuePasa is a Go-based WhatsApp bot platform that exposes HTTP APIs
    for WhatsApp messaging integration
  license:
    name: GNU Affero General Public License v3.0
    url: https://github.com/nocodeleaks/quepasa/blob/main/LICENSE.md
  termsOfService: https://github.com/nocodeleaks/quepasa
  title: QuePasa WhatsApp API
  version: 4.0.0
paths:
  /account:
    post:
      consumes:
      - application/json
      description: Create, update, or manage user accounts (master access required)
      parameters:
      - description: Account request
        in: body
        name: request
        schema:
          properties:
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Manage user accounts
      tags:
      - Application
  /chat/archive:
    post:
      consumes:
      - application/json
      description: Archives or unarchives a WhatsApp chat. Archiving also unpins the
        chat automatically.
      parameters:
      - description: Chat archive request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ChatArchiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Archive or unarchive chat
      tags:
      - Chat
  /chat/markread:
    post:
      consumes:
      - application/json
      description: Marks a WhatsApp chat as read (removes unread badge)
      parameters:
      - description: Chat read request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ChatReadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Mark chat as read
      tags:
      - Chat
  /chat/markunread:
    post:
      consumes:
      - application/json
      description: Marks a WhatsApp chat as unread (shows unread badge)
      parameters:
      - description: Chat unread request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ChatReadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Mark chat as unread
      tags:
      - Chat
  /chat/presence:
    post:
      consumes:
      - application/json
      description: Controls typing indicators and chat presence in WhatsApp conversations
      parameters:
      - description: Chat presence request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ChatPresenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Control chat presence
      tags:
      - Chat
  /command:
    get:
      consumes:
      - application/json
      description: Execute control commands for the bot server (start, stop, restart)
        or toggle settings (groups, broadcasts, readreceipts, readupdate, calls, debug)
      parameters:
      - description: Command action
        enum:
        - start
        - stop
        - restart
        - groups
        - broadcasts
        - readreceipts
        - readupdate
        - calls
        - debug
        in: query
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Execute bot commands
      tags:
      - Bot
  /contact/search:
    post:
      consumes:
      - application/json
      description: Search WhatsApp contacts by name, phone, and other filters
      parameters:
      - description: Search criteria
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.QpContactsSearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpContactsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Search contacts
      tags:
      - Contacts
  /contacts:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all WhatsApp contacts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpContactsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Get contacts
      tags:
      - Contacts
  /download:
    get:
      description: Downloads media files (images, videos, documents) from WhatsApp
        messages
      parameters:
      - description: Message ID (query parameter alternate)
        in: query
        name: messageid
        type: string
      - description: Use cached content
        in: query
        name: cache
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Media file
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Download media
      tags:
      - Download
  /edit:
    put:
      consumes:
      - application/json
      description: Edits the content of an existing message by its ID
      parameters:
      - description: Message edit request
        in: body
        name: request
        required: true
        schema:
          properties:
            content:
              type: string
            messageId:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Edit message
      tags:
      - Message
  /environment:
    get:
      consumes:
      - application/json
      description: Get all environment variables and configurations (master key required
        for full access, preview available without auth)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.EnvironmentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.EnvironmentResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.EnvironmentResponse'
      security:
      - ApiKeyAuth: []
      summary: Get environment settings
      tags:
      - Environment
  /groups/create:
    post:
      consumes:
      - application/json
      description: Creates a new WhatsApp group with specified title and participants
      parameters:
      - description: Group creation request
        in: body
        name: request
        required: true
        schema:
          properties:
            participants:
              items:
                type: string
              type: array
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpSingleGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new group
      tags:
      - Groups
  /groups/description:
    put:
      consumes:
      - application/json
      description: Updates the topic/description of a specific WhatsApp group
      parameters:
      - description: Group topic update request
        in: body
        name: request
        required: true
        schema:
          properties:
            group_jid:
              type: string
            topic:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpSingleGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Set group topic
      tags:
      - Groups
  /groups/get:
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific WhatsApp group
      parameters:
      - description: Group ID
        in: query
        name: groupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpSingleGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Get group information
      tags:
      - Groups
  /groups/getall:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all WhatsApp groups that the bot is currently
        a member of
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpGroupsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all groups
      tags:
      - Groups
  /groups/leave:
    post:
      consumes:
      - application/json
      description: Leave a specific WhatsApp group
      parameters:
      - description: Leave group request
        in: body
        name: request
        required: true
        schema:
          properties:
            chatId:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Leave group
      tags:
      - Groups
  /groups/name:
    put:
      consumes:
      - application/json
      description: Updates the name of a specific WhatsApp group
      parameters:
      - description: Group name update request
        in: body
        name: request
        required: true
        schema:
          properties:
            group_jid:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpSingleGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Set group name
      tags:
      - Groups
  /groups/participants:
    put:
      consumes:
      - application/json
      description: Add, remove, promote, or demote participants in a WhatsApp group
      parameters:
      - description: Participants update request
        in: body
        name: request
        required: true
        schema:
          properties:
            action:
              type: string
            group_jid:
              type: string
            participants:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpParticipantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Update group participants
      tags:
      - Groups
  /groups/photo:
    put:
      consumes:
      - application/json
      description: Updates or removes the photo of a specific WhatsApp group
      parameters:
      - description: Group photo update request
        in: body
        name: request
        required: true
        schema:
          properties:
            group_jid:
              type: string
            remove_img:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpSingleGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Set/Remove group photo
      tags:
      - Groups
  /groups/requests:
    get:
      consumes:
      - application/json
      description: Get, approve, or reject join requests for WhatsApp groups
      parameters:
      - description: Membership request
        in: body
        name: request
        required: true
        schema:
          properties:
            action:
              type: string
            group_jid:
              type: string
            participants:
              items:
                type: string
              type: array
          type: object
      - description: Group JID (for GET requests)
        in: query
        name: group_jid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpRequestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Handle group join requests
      tags:
      - Groups
    post:
      consumes:
      - application/json
      description: Get, approve, or reject join requests for WhatsApp groups
      parameters:
      - description: Membership request
        in: body
        name: request
        required: true
        schema:
          properties:
            action:
              type: string
            group_jid:
              type: string
            participants:
              items:
                type: string
              type: array
          type: object
      - description: Group JID (for GET requests)
        in: query
        name: group_jid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpRequestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Handle group join requests
      tags:
      - Groups
...
  /info:
    delete:
      consumes:
      - application/json
      description: Delete bot/server information and settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InformationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete bot information
      tags:
      - Information
    get:
      consumes:
      - application/json
      description: Get bot/server information and settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InformationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Get bot information
      tags:
      - Information
    patch:
      consumes:
      - application/json
      description: Update bot/server information and settings
      parameters:
      - description: Settings update
        in: body
        name: request
        schema:
          properties:
            settings:
              type: object
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InformationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Update bot information
      tags:
      - Information
    post:
      consumes:
      - application/json
      description: Create a new bot/server with configuration before QR code scanning
      parameters:
      - description: Server creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.InfoCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Server updated
          schema:
            $ref: '#/definitions/api.InformationResponse'
        "201":
          description: Server created
          schema:
            $ref: '#/definitions/api.InformationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Create bot configuration
      tags:
      - Information
  /invite:
    get:
      consumes:
      - application/json
      description: Generates an invite link for a specific WhatsApp group
      parameters:
      - description: Chat ID (query parameter)
        in: query
        name: chatid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpInviteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate group invite link
      tags:
      - Groups
  /invite/{chatid}:
    get:
      consumes:
      - application/json
      description: Generates an invite link for a specific WhatsApp group
      parameters:
      - description: Chat ID (path parameter)
        in: path
        name: chatid
        type: string
      - description: Chat ID (query parameter)
        in: query
        name: chatid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpInviteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate group invite link
      tags:
      - Groups
  /isonwhatsapp:
    post:
      consumes:
      - application/json
      description: Checks if provided phone numbers are registered on WhatsApp
      parameters:
      - description: Phone numbers to check
        in: body
        name: request
        required: true
        schema:
          properties:
            phones:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpIsOnWhatsappResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Check WhatsApp registration
      tags:
      - Contacts
  /message/{messageid}:
    delete:
      consumes:
      - application/json
      description: Revokes or deletes a specific message by its ID
      parameters:
      - description: Message ID
        in: path
        name: messageid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Revoke message
      tags:
      - Message
    get:
      consumes:
      - application/json
      description: Retrieves a specific message by its ID
      parameters:
      - description: Message ID
        in: path
        name: messageid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Get message
      tags:
      - Message
  /paircode:
    get:
      description: Generates a pairing code for WhatsApp authentication using phone
        number
      parameters:
      - description: Phone number for pairing
        in: query
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate pairing code
      tags:
      - Connection
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-QUEPASA-TOKEN
    type: apiKey
swagger: "2.0"